+++ packages/ui/v2/core/Shell.tsx
@@ -1,13 +1,13 @@
 import type { User } from "@prisma/client";
+import noop from "lodash/noop";
 import { signOut, useSession } from "next-auth/react";
 import Link from "next/link";
 import { NextRouter, useRouter } from "next/router";
-import React, { Fragment, ReactNode, useEffect, useState } from "react";
+import React, { Dispatch, Fragment, ReactNode, SetStateAction, useEffect, useState } from "react";
 import { Toaster } from "react-hot-toast";
 
 import dayjs from "@calcom/dayjs";
 import { useIsEmbed } from "@calcom/embed-core/embed-iframe";
-import LicenseBanner from "@calcom/features/ee/common/components/LicenseBanner";
 import TrialBanner from "@calcom/features/ee/common/components/TrialBanner";
 import ImpersonatingBanner from "@calcom/features/ee/impersonation/components/ImpersonatingBanner";
 import HelpMenuItem from "@calcom/features/ee/support/components/HelpMenuItem";
@@ -35,8 +35,6 @@ import { KBarContent, KBarRoot, KBarTrigger } from "../../Kbar";
 import Logo from "../../Logo";
 // TODO: re-introduce in 2.1 import Tips from "../modules/tips/Tips";
 import HeadSeo from "./head-seo";
-import MobileSettingsContainer from "./navigation/MobileSettingsContainer";
-import SettingsSidebarContainer from "./navigation/SettingsSidebarContainer";
 import { SkeletonText } from "./skeleton";
 
 /* TODO: Migate this */
@@ -138,15 +136,7 @@ const Layout = (props: LayoutProps) => {
       </div>
 
       <div className="flex h-screen overflow-hidden" data-testid="dashboard-shell">
-        {props.settingsSection ? (
-          <div className="hidden lg:block">
-            <SettingsSidebarContainer />
-          </div>
-        ) : props.SidebarContainer ? (
-          props.SidebarContainer
-        ) : (
-          <SideBarContainer />
-        )}
+        {props.SidebarContainer || <SideBarContainer />}
         <div className="flex w-0 flex-1 flex-col overflow-hidden">
           <UserV2OptInBanner />
           <ImpersonatingBanner />
@@ -157,6 +147,8 @@ const Layout = (props: LayoutProps) => {
   );
 };
 
+type DrawerState = [isOpen: boolean, setDrawerOpen: Dispatch<SetStateAction<boolean>>];
+
 type LayoutProps = {
   centered?: boolean;
   title?: string;
@@ -165,7 +157,11 @@ type LayoutProps = {
   children: ReactNode;
   CTA?: ReactNode;
   large?: boolean;
+  SettingsSidebarContainer?: ReactNode;
+  MobileNavigationContainer?: ReactNode;
   SidebarContainer?: ReactNode;
+  TopNavContainer?: ReactNode;
+  drawerState?: DrawerState;
   HeadingLeftIcon?: ReactNode;
   backPath?: string; // renders back button to specified path
   // use when content needs to expand with flex
@@ -690,17 +686,20 @@ export function ShellMain(props: LayoutProps) {
   );
 }
 
-function MainContainer(props: LayoutProps) {
-  const [sideContainerOpen, setSideContainerOpen] = useState(false);
+const SettingsSidebarContainerDefault = () => null;
+
+function MainContainer({
+  SettingsSidebarContainer: SettingsSidebarContainerProp = <SettingsSidebarContainerDefault />,
+  MobileNavigationContainer: MobileNavigationContainerProp = <MobileNavigationContainer />,
+  TopNavContainer: TopNavContainerProp = <TopNavContainer />,
+  ...props
+}: LayoutProps) {
+  const [sideContainerOpen, setSideContainerOpen] = props.drawerState || [false, noop];
 
   return (
     <main className="relative z-0 flex flex-1 flex-col overflow-y-auto bg-white focus:outline-none ">
       {/* show top navigation for md and smaller (tablet and phones) */}
-      {props.settingsSection ? (
-        <MobileSettingsContainer onSideContainerOpen={() => setSideContainerOpen(!sideContainerOpen)} />
-      ) : (
-        <TopNavContainer />
-      )}
+      {TopNavContainerProp}
       {/* The following is used for settings navigation on medium and smaller screens */}
       <div
         className={classNames(
@@ -716,14 +715,14 @@ function MainContainer(props: LayoutProps) {
           "absolute inset-y-0 z-50 m-0 h-screen w-56 transform border-gray-100 bg-gray-50 transition duration-200 ease-in-out",
           sideContainerOpen ? "translate-x-0" : "-translate-x-full"
         )}>
-        <SettingsSidebarContainer />
+        {SettingsSidebarContainerProp}
       </div>
       <div className="px-4 py-2 lg:py-8 lg:px-12">
         <ErrorBoundary>
           {!props.withoutMain ? <ShellMain {...props}>{props.children}</ShellMain> : props.children}
         </ErrorBoundary>
         {/* show bottom navigation for md and smaller (tablet and phones) */}
-        {!props.settingsSection && <MobileNavigationContainer />}
+        {MobileNavigationContainerProp}
         {/* <LicenseBanner /> */}
       </div>
     </main>
